Avoid hard-coding TF resource renames for Chalice
