+++ scripts/prepare_lambda_deployment.py
@@ -3,14 +3,8 @@ from argparse import (
 )
 import json
 import logging
-from pathlib import (
-    Path,
-)
 import shutil
 import sys
-from typing import (
-    TypeVar,
-)
 
 from azul import (
     config,
@@ -22,11 +16,8 @@ from azul.logging import (
     configure_script_logging,
 )
 from azul.terraform import (
+    chalice,
     populate_tags,
-)
-from azul.types import (
-    AnyJSON,
-    JSON,
 )
 
 log = logging.getLogger(__name__)
@@ -76,141 +67,31 @@ def transform_tf(input_json):
     return input_json
 
 
-def patch_resource_names(tf_config: JSON) -> JSON:
-    """
-    Some Chalice-generated resources have named ending in `-event`. The
-    dash prevents generation of a fully qualified resource name (i.e.,
-    with `config.qualified_resource_name`. This function returns
-    Terraform configuration with names and references updated to
-    remove the trailing `-event`.
-
-    >>> from azul.doctests import assert_json
-    >>> assert_json(patch_resource_names({
-    ...     "resource": {
-    ...         "aws_cloudwatch_event_rule": {
-    ...             "indexercachehealth-event": {  # patch
-    ...                 "name": "indexercachehealth-event"  # leave
-    ...             }
-    ...         },
-    ...         "aws_cloudwatch_event_target": {
-    ...             "indexercachehealth-event": {  # patch
-    ...                 "rule": "${aws_cloudwatch_event_rule.indexercachehealth-event.name}",  # patch
-    ...                 "target_id": "indexercachehealth-event",  # leave
-    ...                 "arn": "${aws_lambda_function.indexercachehealth.arn}"
-    ...             }
-    ...         },
-    ...         "aws_lambda_permission": {
-    ...             "indexercachehealth-event": {  # patch
-    ...                 "function_name": "azul-indexer-prod-indexercachehealth",
-    ...                 "source_arn": "${aws_cloudwatch_event_rule.indexercachehealth-event.arn}"  # patch
-    ...             }
-    ...         },
-    ...         "aws_lambda_event_source_mapping": {
-    ...             "contribute-sqs-event-source": {
-    ...                 "batch_size": 1
-    ...             }
-    ...         }
-    ...     }
-    ... }))
-    {
-        "resource": {
-            "aws_cloudwatch_event_rule": {
-                "indexercachehealth": {
-                    "name": "indexercachehealth-event"
-                }
-            },
-            "aws_cloudwatch_event_target": {
-                "indexercachehealth": {
-                    "rule": "${aws_cloudwatch_event_rule.indexercachehealth.name}",
-                    "target_id": "indexercachehealth-event",
-                    "arn": "${aws_lambda_function.indexercachehealth.arn}"
-                }
-            },
-            "aws_lambda_permission": {
-                "indexercachehealth": {
-                    "function_name": "azul-indexer-prod-indexercachehealth",
-                    "source_arn": "${aws_cloudwatch_event_rule.indexercachehealth.arn}"
-                }
-            },
-            "aws_lambda_event_source_mapping": {
-                "contribute-sqs-event-source": {
-                    "batch_size": 1
-                }
-            }
-        }
-    }
-    """
-    suffix = '-event'
-
-    mapping = {}
-    for resource_type, resources in tf_config['resource'].items():
-        for name in resources:
-            if name.endswith(suffix):
-                new_name = name[:-len(suffix)]
-                mapping[resource_type, name] = new_name
-
-    tf_config = {
-        block_name: {
-            resource_type: {
-                mapping.get((resource_type, name), name): resource
-                for name, resource in resources.items()
-            }
-            for resource_type, resources in block.items()
-        } if block_name == 'resource' else block
-        for block_name, block in tf_config.items()
-    }
-
-    def ref(resource_type, name):
-        return '${' + resource_type + '.' + name + '.'
-
-    ref_map = {
-        ref(resource_type, name): ref(resource_type, new_name)
-        for (resource_type, name), new_name in mapping.items()
-    }
-
-    def patch_refs(v: U) -> U:
-        if isinstance(v, dict):
-            return {k: patch_refs(v) for k, v in v.items()}
-        elif isinstance(v, str):
-            for old_ref, new_ref in ref_map.items():
-                if old_ref in v:
-                    return v.replace(old_ref, new_ref)
-            return v
-        else:
-            return v
-
-    return patch_refs(tf_config)
-
-
-U = TypeVar('U', bound=AnyJSON)
-
-
 def main(argv):
     parser = ArgumentParser(
-        description='Prepare the Terraform config generated by `chalice package'
-                    '--pkg-format terraform` and copy it into the terraform/ '
-                    'directory.'
+        description='Prepare the Terraform config generated by '
+                    '`chalice package --pkg-format terraform` '
+                    'and copy it into the terraform/ directory.'
     )
-    parser.add_argument('lambda_name', help='the lambda of the config that will be '
-                                            'transformed and copied')
+    parser.add_argument('lambda_name',
+                        help='The name of the Lambda function to prepare.')
     options = parser.parse_args(argv)
-    source_dir = Path(config.project_root) / 'lambdas' / options.lambda_name / '.chalice' / 'terraform'
-    output_dir = Path(config.project_root) / 'terraform' / options.lambda_name
-    output_dir.mkdir(exist_ok=True)
+    lambda_name = options.lambda_name
+    src_dir = chalice.package_dir(lambda_name)
+    dst_dir = chalice.module_dir(lambda_name)
+    dst_dir.mkdir(exist_ok=True)
 
-    deployment_src = source_dir / 'deployment.zip'
-    deployment_dst = output_dir / 'deployment.zip'
-    log.info('Copying %s to %s', deployment_src, deployment_dst)
-    shutil.copyfile(deployment_src, deployment_dst)
+    args = [dir / chalice.package_zip_name for dir in (src_dir, dst_dir)]
+    log.info('Copying %s to %s', *args)
+    shutil.copyfile(*args)
 
-    tf_src = source_dir / 'chalice.tf.json'
-    tf_dst = output_dir / 'chalice.tf.json'
-    log.info('Transforming %s to %s', tf_src, tf_dst)
-    with open(tf_src, 'r') as f:
-        output_json = json.load(f)
-    output_json = populate_tags(patch_resource_names(transform_tf(output_json)))
-    with write_file_atomically(tf_dst) as f:
-        json.dump(output_json, f, indent=4)
+    src_tf, dst_tf = [dir / chalice.tf_config_name for dir in (src_dir, dst_dir)]
+    log.info('Transforming %s to %s', src_tf, dst_tf)
+    with open(src_tf) as f:
+        tf_config = json.load(f)
+    tf_config = populate_tags(chalice.patch_resource_names(transform_tf(tf_config)))
+    with write_file_atomically(dst_tf) as f:
+        json.dump(tf_config, f, indent=4)
 
 
 if __name__ == '__main__':
